chore: refactor components
